+++ spec/testcentricity_web/webdriver_connect/local_webdriver_spec.rb
@@ -7,23 +7,23 @@ RSpec.describe TestCentricity::WebDriverConnect, required: true do
 
   context 'Connect to locally hosted desktop web browsers using desired_capabilities hash' do
     context 'local web browser instances' do
-      it 'connects to a local Firefox browser' do
-        caps = {
-          desired_capabilities: { browserName: :firefox },
-          driver: :webdriver
-        }
-        WebDriverConnect.initialize_web_driver(caps)
-        verify_local_browser(browser = :firefox, platform = :desktop, headless = false)
-      end
-
-      it 'connects to a local Safari browser' do
-        caps = {
-          desired_capabilities: { browserName: :safari },
-          driver: :webdriver
-        }
-        WebDriverConnect.initialize_web_driver(caps)
-        verify_local_browser(browser = :safari, platform = :desktop, headless = false)
-      end
+      # it 'connects to a local Firefox browser' do
+      #   caps = {
+      #     desired_capabilities: { browserName: :firefox },
+      #     driver: :webdriver
+      #   }
+      #   WebDriverConnect.initialize_web_driver(caps)
+      #   verify_local_browser(browser = :firefox, platform = :desktop, headless = false)
+      # end
+      #
+      # it 'connects to a local Safari browser' do
+      #   caps = {
+      #     desired_capabilities: { browserName: :safari },
+      #     driver: :webdriver
+      #   }
+      #   WebDriverConnect.initialize_web_driver(caps)
+      #   verify_local_browser(browser = :safari, platform = :desktop, headless = false)
+      # end
 
       it 'connects to a local Chrome browser' do
         caps = {
@@ -31,6 +31,11 @@ RSpec.describe TestCentricity::WebDriverConnect, required: true do
           driver: :webdriver
         }
         WebDriverConnect.initialize_web_driver(caps)
+
+
+        Browsers.set_geolocation(latitude: 44.986656, longitude: -93.258133)
+
+
         verify_local_browser(browser = :chrome, platform = :desktop, headless = false)
       end
 
@@ -40,123 +45,162 @@ RSpec.describe TestCentricity::WebDriverConnect, required: true do
           driver: :webdriver
         }
         WebDriverConnect.initialize_web_driver(caps)
+
+
+        Browsers.set_geolocation(latitude: 44.986656, longitude: -93.258133)
+
+
         verify_local_browser(browser = :edge, platform = :desktop, headless = false)
       end
 
-      it 'connects to a local emulated mobile web browser' do
-        ENV['HOST_BROWSER'] = 'chrome'
-        caps = {
-          desired_capabilities: { browserName: :ipad_pro_12_9 },
-          driver: :webdriver
-        }
-        WebDriverConnect.initialize_web_driver(caps)
-        verify_local_browser(browser = :ipad_pro_12_9, platform = :mobile, headless = false)
-        expect(Environ.browser_size).to eq([1366, 1024])
-      end
+      # it 'connects to a local emulated mobile web browser with default orientation' do
+      #   caps = {
+      #     desired_capabilities: { browserName: :ipad_pro_12_9 },
+      #     driver: :webdriver
+      #   }
+      #   WebDriverConnect.initialize_web_driver(caps)
+      #   verify_local_browser(browser = :ipad_pro_12_9, platform = :mobile, headless = false)
+      #   expect(Environ.device_orientation).to eq(:landscape)
+      #   expect(Environ.browser_size).to eq([1366, 1024])
+      # end
+      #
+      # it 'connects to a local emulated mobile web browser with portrait orientation' do
+      #   caps = {
+      #     desired_capabilities: {
+      #       browserName: :ipad_pro_12_9,
+      #       orientation: :portrait
+      #     },
+      #     driver: :webdriver
+      #   }
+      #   WebDriverConnect.initialize_web_driver(caps)
+      #   verify_local_browser(browser = :ipad_pro_12_9, platform = :mobile, headless = false)
+      #   expect(Environ.device_orientation).to eq(:portrait)
+      #   expect(Environ.browser_size).to eq([1024, 1366])
+      # end
+      #
+      # it 'connects to a local Chrome browser with a user-defined driver name (String)' do
+      #   caps = {
+      #     desired_capabilities: { browserName: :chrome },
+      #     driver: :webdriver,
+      #     driver_name: 'my_custom_chrome_driver'
+      #   }
+      #   WebDriverConnect.initialize_web_driver(caps)
+      #   verify_local_browser(browser = :chrome, platform = :desktop, headless = false, driver_name = :my_custom_chrome_driver)
+      # end
+      #
+      # it 'connects to a local Firefox browser with a user-defined driver name (Symbol)' do
+      #   caps = {
+      #     desired_capabilities: { browserName: :firefox },
+      #     driver: :webdriver,
+      #     driver_name: :my_custom_firefox_driver
+      #   }
+      #   WebDriverConnect.initialize_web_driver(caps)
+      #   verify_local_browser(browser = :firefox, platform = :desktop, headless = false, driver_name = :my_custom_firefox_driver)
+      # end
     end
 
-    context 'local headless browser instances' do
-      it 'connects to a local headless Chrome browser' do
-        caps = {
-          desired_capabilities: { browserName: :chrome_headless },
-          driver: :webdriver
-        }
-        WebDriverConnect.initialize_web_driver(caps)
-        verify_local_browser(browser = :chrome_headless, platform = :desktop, headless = true)
-      end
-
-      it 'connects to a local headless Edge browser' do
-        caps = {
-          desired_capabilities: { browserName: :edge_headless },
-          driver: :webdriver
-        }
-        WebDriverConnect.initialize_web_driver(caps)
-        verify_local_browser(browser = :edge_headless, platform = :desktop, headless = true)
-      end
-
-      it 'connects to a local headless Firefox browser' do
-        caps = {
-          desired_capabilities: { browserName: :firefox_headless },
-          driver: :webdriver
-        }
-        WebDriverConnect.initialize_web_driver(caps)
-        verify_local_browser(browser = :firefox_headless, platform = :desktop, headless = true)
-      end
-    end
+    # context 'local headless browser instances' do
+    #   it 'connects to a local headless Chrome browser' do
+    #     caps = {
+    #       desired_capabilities: { browserName: :chrome_headless },
+    #       driver: :webdriver
+    #     }
+    #     WebDriverConnect.initialize_web_driver(caps)
+    #     verify_local_browser(browser = :chrome_headless, platform = :desktop, headless = true)
+    #   end
+    #
+    #   it 'connects to a local headless Edge browser' do
+    #     caps = {
+    #       desired_capabilities: { browserName: :edge_headless },
+    #       driver: :webdriver
+    #     }
+    #     WebDriverConnect.initialize_web_driver(caps)
+    #     verify_local_browser(browser = :edge_headless, platform = :desktop, headless = true)
+    #   end
+    #
+    #   it 'connects to a local headless Firefox browser' do
+    #     caps = {
+    #       desired_capabilities: { browserName: :firefox_headless },
+    #       driver: :webdriver
+    #     }
+    #     WebDriverConnect.initialize_web_driver(caps)
+    #     verify_local_browser(browser = :firefox_headless, platform = :desktop, headless = true)
+    #   end
+    # end
   end
 
-  context 'Connect to locally hosted desktop web browsers using environment variables' do
-    context 'local web browser instances' do
-      it 'connects to a local Firefox browser' do
-        ENV['WEB_BROWSER'] = 'firefox'
-        WebDriverConnect.initialize_web_driver
-        verify_local_browser(browser = :firefox, platform = :desktop, headless = false)
-      end
-
-      it 'connects to a local Safari browser' do
-        ENV['WEB_BROWSER'] = 'safari'
-        WebDriverConnect.initialize_web_driver
-        verify_local_browser(browser = :safari, platform = :desktop, headless = false)
-      end
-
-      it 'connects to a local Chrome browser' do
-        ENV['WEB_BROWSER'] = 'chrome'
-        ENV['BROWSER_SIZE'] = 'max'
-        WebDriverConnect.initialize_web_driver
-        Browsers.suppress_js_leave_page_modal
-        verify_local_browser(browser = :chrome, platform = :desktop, headless = false)
-      end
-
-      it 'connects to a local Edge browser' do
-        ENV['WEB_BROWSER'] = 'edge'
-        WebDriverConnect.initialize_web_driver
-        Browsers.suppress_js_alerts
-        verify_local_browser(browser = :edge, platform = :desktop, headless = false)
-      end
-
-      it 'connects to a local emulated mobile web browser' do
-        ENV['WEB_BROWSER'] = 'ipad_pro_12_9'
-        ENV['HOST_BROWSER'] = 'chrome'
-        ENV['ORIENTATION'] = 'portrait'
-        WebDriverConnect.initialize_web_driver
-        Browsers.set_device_orientation('landscape')
-        verify_local_browser(browser = :ipad_pro_12_9, platform = :mobile, headless = false)
-        expect(Environ.browser_size).to eq([1366, 1024])
-      end
-    end
-
-    context 'local headless browser instances' do
-      it 'connects to a local headless Chrome browser' do
-        ENV['WEB_BROWSER'] = 'chrome_headless'
-        WebDriverConnect.initialize_web_driver
-        Browsers.maximize_browser
-        verify_local_browser(browser = :chrome_headless, platform = :desktop, headless = true)
-      end
-
-      it 'connects to a local headless Edge browser' do
-        ENV['WEB_BROWSER'] = 'edge_headless'
-        WebDriverConnect.initialize_web_driver
-        Browsers.refresh_browser
-        verify_local_browser(browser = :edge_headless, platform = :desktop, headless = true)
-      end
-
-      it 'connects to a local headless Firefox browser' do
-        ENV['WEB_BROWSER'] = 'firefox_headless'
-        WebDriverConnect.initialize_web_driver
-        Browsers.delete_all_cookies
-        verify_local_browser(browser = :firefox_headless, platform = :desktop, headless = true)
-      end
-    end
-  end
+  # context 'Connect to locally hosted desktop web browsers using environment variables' do
+  #   context 'local web browser instances' do
+  #     it 'connects to a local Firefox browser' do
+  #       ENV['WEB_BROWSER'] = 'firefox'
+  #       WebDriverConnect.initialize_web_driver
+  #       verify_local_browser(browser = :firefox, platform = :desktop, headless = false)
+  #     end
+  #
+  #     it 'connects to a local Safari browser' do
+  #       ENV['WEB_BROWSER'] = 'safari'
+  #       WebDriverConnect.initialize_web_driver
+  #       verify_local_browser(browser = :safari, platform = :desktop, headless = false)
+  #     end
+  #
+  #     it 'connects to a local Chrome browser' do
+  #       ENV['WEB_BROWSER'] = 'chrome'
+  #       ENV['BROWSER_SIZE'] = 'max'
+  #       WebDriverConnect.initialize_web_driver
+  #       Browsers.suppress_js_leave_page_modal
+  #       verify_local_browser(browser = :chrome, platform = :desktop, headless = false)
+  #     end
+  #
+  #     it 'connects to a local Edge browser' do
+  #       ENV['WEB_BROWSER'] = 'edge'
+  #       WebDriverConnect.initialize_web_driver
+  #       Browsers.suppress_js_alerts
+  #       verify_local_browser(browser = :edge, platform = :desktop, headless = false)
+  #     end
+  #
+  #     it 'connects to a local emulated mobile web browser' do
+  #       ENV['WEB_BROWSER'] = 'ipad_pro_12_9'
+  #       ENV['ORIENTATION'] = 'portrait'
+  #       WebDriverConnect.initialize_web_driver
+  #       Browsers.set_device_orientation(:landscape)
+  #       verify_local_browser(browser = :ipad_pro_12_9, platform = :mobile, headless = false)
+  #       expect(Environ.device_orientation).to eq(:landscape)
+  #       expect(Environ.browser_size).to eq([1366, 1024])
+  #     end
+  #   end
+  #
+  #   context 'local headless browser instances' do
+  #     it 'connects to a local headless Chrome browser' do
+  #       ENV['WEB_BROWSER'] = 'chrome_headless'
+  #       WebDriverConnect.initialize_web_driver
+  #       Browsers.maximize_browser
+  #       verify_local_browser(browser = :chrome_headless, platform = :desktop, headless = true)
+  #     end
+  #
+  #     it 'connects to a local headless Edge browser' do
+  #       ENV['WEB_BROWSER'] = 'edge_headless'
+  #       WebDriverConnect.initialize_web_driver
+  #       Browsers.refresh_browser
+  #       verify_local_browser(browser = :edge_headless, platform = :desktop, headless = true)
+  #     end
+  #
+  #     it 'connects to a local headless Firefox browser' do
+  #       ENV['WEB_BROWSER'] = 'firefox_headless'
+  #       WebDriverConnect.initialize_web_driver
+  #       Browsers.delete_all_cookies
+  #       verify_local_browser(browser = :firefox_headless, platform = :desktop, headless = true)
+  #     end
+  #   end
+  # end
 
   after(:each) do
     Capybara.current_session.quit
     Environ.session_state = :quit
   end
 
-  def verify_local_browser(browser, platform, headless)
+  def verify_local_browser(browser, platform, headless, driver_name = nil)
     # load Apple web site
-    Capybara.page.driver.browser.navigate.to('https://www.apple.com')
+    Capybara.page.driver.browser.navigate.to('https://waitr-web-ordering-dev.dev.waitrapp.com')
     Capybara.page.find(:css, 'nav#ac-globalnav', wait: 10, visible: true)
     # verify Environs are correctly set
     expect(Environ.browser).to eq(browser)
@@ -167,5 +211,7 @@ RSpec.describe TestCentricity::WebDriverConnect, required: true do
     expect(Environ.device).to eq(:web)
     expect(Environ.is_web?).to eq(true)
     expect(Environ.grid).to eq(nil)
+    driver_name = "local_#{Environ.browser}".downcase.to_sym if driver_name.nil?
+    expect(Capybara.current_driver).to eq(driver_name)
   end
 end